[kimchi][proof] apply move of proof data struct
