+++ kimchi/src/bench.rs
@@ -4,7 +4,7 @@ use crate::{
         polynomials::generic::GenericGateSpec,
         wires::{Wire, COLUMNS},
     },
-    prover::ProverProof,
+    proof::ProverProof,
     prover_index::{testing::new_index_for_test, ProverIndex},
     verifier::batch_verify,
     verifier_index::VerifierIndex,
