+++ kimchi/src/circuits/polynomials/chacha.rs
@@ -505,9 +505,9 @@ mod tests {
         circuits::{
             expr::{Column, Constants, PolishToken},
             lookup::lookups::LookupInfo,
-            scalars::{LookupEvaluations, ProofEvaluations},
             wires::*,
         },
+        proof::{LookupEvaluations, ProofEvaluations},
     };
     use ark_ff::UniformRand;
     use ark_poly::{EvaluationDomain, Radix2EvaluationDomain as D};
