+++ kimchi/src/tests/ec.rs
@@ -3,7 +3,7 @@ use crate::{
         gate::{CircuitGate, GateType},
         wires::*,
     },
-    prover::ProverProof,
+    proof::ProverProof,
     prover_index::testing::new_index_for_test,
     verifier::verify,
 };
