+++ kimchi/src/tests/generic.rs
@@ -1,6 +1,6 @@
 use crate::circuits::polynomials::generic::testing::{create_circuit, fill_in_witness};
 use crate::circuits::{gate::CircuitGate, wires::COLUMNS};
-use crate::prover::ProverProof;
+use crate::proof::ProverProof;
 use crate::prover_index::testing::new_index_for_test;
 use crate::verifier::verify;
 use ark_ff::{UniformRand, Zero};
